+++ src/utility.jl
@@ -1009,10 +1009,12 @@ View of the i-th row(s) or j-th column(s) of a matrix `X`,
 or of the i-th element(s) of vector `x`.
 """ 
 vrow(X, i) = view(X, i, :)
+vrow(X::DataFrame, i) = view(Matrix(X), i, :)
 
 vrow(x::Vector, i) = view(x, i)
 
 vcol(X, j) = view(X, :, j)
+vcol(X::DataFrame, j) = view(Matrix(X), :, j)
 
 vcol(x::Vector, i) = view(x, i)